# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ImpactOS API",
    "description": "Impact & Scenario Intelligence System for Strategic Gears.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/workspaces": {
      "get": {
        "tags": [
          "workspaces"
        ],
        "summary": "List Workspaces",
        "operationId": "list_workspaces_v1_workspaces_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceListResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "workspaces"
        ],
        "summary": "Create Workspace",
        "operationId": "create_workspace_v1_workspaces_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWorkspaceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/workspaces/{workspace_id}/scenarios": {
      "get": {
        "tags": [
          "scenarios"
        ],
        "summary": "List Scenarios",
        "description": "B-9: List scenarios for a workspace (latest version per spec, paginated).",
        "operationId": "list_scenarios_v1_workspaces__workspace_id__scenarios_get",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Workspace Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScenarioListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "scenarios"
        ],
        "summary": "Create Scenario",
        "description": "Create a new scenario (version 1).",
        "operationId": "create_scenario_v1_workspaces__workspace_id__scenarios_post",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Workspace Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateScenarioRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateScenarioResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/workspaces/{workspace_id}/scenarios/{scenario_id}": {
      "get": {
        "tags": [
          "scenarios"
        ],
        "summary": "Get Scenario Detail",
        "description": "B-10: Get full scenario detail (latest version, workspace-scoped).",
        "operationId": "get_scenario_detail_v1_workspaces__workspace_id__scenarios__scenario_id__get",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Workspace Id"
            }
          },
          {
            "name": "scenario_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Scenario Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScenarioDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/workspaces/{workspace_id}/scenarios/{scenario_id}/compile": {
      "post": {
        "tags": [
          "scenarios"
        ],
        "summary": "Compile Scenario",
        "description": "Compile line items + decisions into shock items.\n\nAccepts EITHER document_id (preferred) OR inline line_items (deprecated).",
        "operationId": "compile_scenario_v1_workspaces__workspace_id__scenarios__scenario_id__compile_post",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Workspace Id"
            }
          },
          {
            "name": "scenario_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Scenario Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/src__api__scenarios__CompileRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Compile Scenario V1 Workspaces  Workspace Id  Scenarios  Scenario Id  Compile Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/workspaces/{workspace_id}/scenarios/{scenario_id}/versions": {
      "get": {
        "tags": [
          "scenarios"
        ],
        "summary": "Get Versions",
        "description": "Get all versions of a scenario.",
        "operationId": "get_versions_v1_workspaces__workspace_id__scenarios__scenario_id__versions_get",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Workspace Id"
            }
          },
          {
            "name": "scenario_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Scenario Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VersionsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/workspaces/{workspace_id}/scenarios/{scenario_id}/lock": {
      "post": {
        "tags": [
          "scenarios"
        ],
        "summary": "Lock Scenario",
        "description": "Lock mappings for governed run \u2014 creates new version.",
        "operationId": "lock_scenario_v1_workspaces__workspace_id__scenarios__scenario_id__lock_post",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Workspace Id"
            }
          },
          {
            "name": "scenario_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Scenario Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LockRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LockResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateScenarioResponse": {
        "properties": {
          "scenario_spec_id": {
            "type": "string",
            "title": "Scenario Spec Id"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "name": {
            "type": "string",
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "scenario_spec_id",
          "version",
          "name"
        ],
        "title": "CreateScenarioResponse"
      },
      "CreateWorkspaceRequest": {
        "properties": {
          "client_name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Client Name"
          },
          "engagement_code": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Engagement Code"
          },
          "classification": {
            "allOf": [
              {
                "$ref": "#/components/schemas/DataClassification"
              }
            ],
            "default": "CONFIDENTIAL"
          },
          "description": {
            "type": "string",
            "maxLength": 2000,
            "title": "Description",
            "default": ""
          },
          "created_by": {
            "type": "string",
            "format": "uuid",
            "title": "Created By"
          }
        },
        "type": "object",
        "required": [
          "client_name",
          "engagement_code",
          "created_by"
        ],
        "title": "CreateWorkspaceRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LockRequest": {
        "properties": {
          "actor": {
            "type": "string",
            "title": "Actor"
          }
        },
        "type": "object",
        "required": [
          "actor"
        ],
        "title": "LockRequest"
      },
...
      "ScenarioListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ScenarioListItem"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total"
        ],
        "title": "ScenarioListResponse"
      },
...
      "WorkspaceListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/WorkspaceResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total"
        ],
        "title": "WorkspaceListResponse"
      },
      "WorkspaceResponse": {
        "properties": {
          "workspace_id": {
            "type": "string",
            "title": "Workspace Id"
          },
          "client_name": {
            "type": "string",
            "title": "Client Name"
          },
          "engagement_code": {
            "type": "string",
            "title": "Engagement Code"
          },
          "classification": {
            "type": "string",
            "title": "Classification"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "created_by": {
            "type": "string",
            "title": "Created By"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "workspace_id",
          "client_name",
          "engagement_code",
          "classification",
          "description",
          "created_by",
          "created_at",
          "updated_at"
        ],
        "title": "WorkspaceResponse"
      },
...
    }
  }
}